moved BaseAST::filename to ModuleSymbol::filename
This change means that we no longer keep track (or have to keep track)
of the filename on every node in the IR, but only on the module
symbols.

This reduces the memory footprint significantly (by nearly 2MB for
jacobi right after the resolve pass).

Reworked handling of error messages related to setting config
variables.  This used to be handled by changing the filename field of
the _cast call to "<command line setting ...>".  It now works by
calling a function _command_line_cast that has a pragma to indicate
special handling is needed for dealing with the runtime line number.

The only gotcha is that when raising an error, the filename reporting
will only work if the BaseAST the error is associated with is in the
IR, i.e., has a path of parentSymbol pointers that can be traced to a
module.

Changed the line numbers in two .goods by one.  It is still correct
since either line can be reported.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14120 3a8e244f-b0f2-452b-bcba-4c88e055c3ca