reimplemented the traversal mechanism
Replaced the function get_ast_children with the macro
AST_CHILDREN_CALL(ast, function, ...) so that the function is called
on every child of ast where the additional arguments are passed as
additional arguments to the function.  For example, collect_asts
collects all of the children ast into a vector and uses a preorder
traversal as follows:

  void collect_asts(BaseAST* ast, Vec<BaseAST*>& asts) {
    asts.add(ast);
    AST_CHILDREN_CALL(ast, collect_asts, asts);
  }

Refactored the formal argument list for collect_asts, etc. so that the
ast is first and the vector is second.

I'm seeing a compilation speedup of 7-8% on jacobi and madness as a
result of this change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14121 3a8e244f-b0f2-452b-bcba-4c88e055c3ca