simplified pragma lookup
Before this change, to determine if a symbol had a pragma, the
symbol's list of pragmas would be checked and the enclosing module's
list would also be checked.  We don't ever store a pragma on the
enclosing module (we used to), so I removed this feature.

Removed an unused pragma.

The overall compile-time speedup is measurable but minor.  I'm seeing
jacobi compile in just over 5 seconds on my desktop.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14123 3a8e244f-b0f2-452b-bcba-4c88e055c3ca