Operator precedence modifications
Made most of the operator precedence changes that we discussed
in our meeting when David was here.  In particular:

- made : bind more tightly than **
- moved bitwise &, ^, | down, as in C
- moved reductions up near the unary operators
- moved ! up with ~
- moved unary + and - below *,/,% which is where the last word
  in our email conversation was

The main thing from our meeting that I did not attempt was to do
something smarter with the bottom end of the table, such as
collaprsing if, for, and forall into a single rule; nor doing
something smart with by, .., and in.  Other than those cases, I think
this contains the changes we discussed.

Also made some cosmetic changes along the way:

- found that we had a token -- !: -- in the parser that is unused
  removed this

- improved the error message for attempting to apply a logical
  operator to an integer -- the old one spoke in terms of the isTrue()
  function.  The new one talks about not being able to apply
  short-circuiting logical operators to integers.  This should
  probably be replicated for other datatypes, but I didn't want to go
  too far afield with this change (the new precedences resulted in
  many instances of this error, which I felt compared to fix).


I updated the spec to indicate the new precedence order, and tried to
unify the names in the "use" column a bit, but probably only with
limited success.  I increased our rationale greatly to describe many
of the cases that we had struggled with and why we ended up where we
did in hopes of leaving ourselves a good trail the next time we
question our choices.  I also added some new tests that show off the
cases that motivated our decisions.

I added a pair of future "error message" tests in which we reveal more
about our internal implementation than we should.

I also updated existing tests that relied on the old precedences.  In
particular:

- several tests that relied on "+ reduce A * B"
     test/arrays/bradc/workarounds/arrayOfSpsArray2.chpl
     test/arrays/bradc/workarounds/arrayOfArray2.chpl
     test/users/shetag/tensor.chpl
     test/puzzles/deitz/puzzle071004.chpl (also added a copy
       capturing what the original puzzle did)
     [all madness tests]
     [several reduction-based sudoku tests]

- several tests that relied on the a & b == 0 precedence
     test/trivial/deitz/test_bmm.chpl
     test/trivial/shannon/compSampler.good
     test/trivial/shannon/compSampler.chpl
     [several old ra tests]

- a few tests that relied on -1:uint
     test/trivial/waynew/int2uint.chpl
     test/studies/590o/wk3/04literals.chpl

- removed unfortunate parenthesis that made me start to question
  where we broke the precedence triangle:
     test/trivial/deitz/test_negate_uint64.chpl
     test/trivial/deitz/test_negate_uint32.chpl
     test/trivial/deitz/test_int32.chpl
     test/trivial/marybeth/casts1.chpl




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14124 3a8e244f-b0f2-452b-bcba-4c88e055c3ca