Modified the sync statement to work more like cobegin/coforall statements
The _EndCount class now has a taskList field to keep track of the tasks that are
created inside a sync statement.  When the locale ID in which the task is
created matches the locale ID of the task list (which is also stored in the
_EndCount class), the task is added to the task list.  The _waitEndCount method
has been modified to first look for tasks in the task list that have not yet
begun executing.  If any such tasks are found, they are executed before blocking
until all remaining tasks finish executing.

This required changing the interfaces for chpl_add_to_task_list() and
chpl_begin(), so the MTA and single-threaded versions were changed to match.
In particular, the ID of the task list's locale is now an argument to
chpl_add_to_task_list(), as well as whether this function should call
chpl_begin().  The latter is needed because now the Chapel compiler generates
calls to chpl_add_to_task_list() instead of to chpl_begin().

The changes to code gen are in expr.cpp.  At compile time, only the wide
reference to the _EndCount class instance is available, not the actual instance
itself.  So, in order to pass in a pointer to the task list as well as the ID
of the task list's locale, the compiler looks at the ID in the wide reference.
If this ID is the same as the ID on which the call is being executed, it
accesses the actual class instance.  Otherwise, it passes NULL and INT32_MIN,
respectively, for the task list and its locale.

A simple test case was added to exhibit the improvement this change makes - the
test case can now be executed with maxThreads set to 1!  Also, the .good file
of another test case was changed, since it contains a line number in
ChapelBase.chpl that is now different.  Unfortunately, the test case in
modules/jplevyak/nostdincs.chpl no longer works, because instances of the
_EndCount class need to be initialized with the ID of the locale where the
instance is created.  I didn't know how best to make this initialization
independent of other standard modules.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14131 3a8e244f-b0f2-452b-bcba-4c88e055c3ca