Add a dynamic cast to an instantiated generic type test
It seems this is the first test that does this. With --no-local, it failed to
get past the C compiler.

Fixed the minor typo/bug that caused it - for wide class dynamic casts, when
doing the codegen for the class ids of the legal classes, the name was used
instead of the cname. In this case it was C(int), looking like a bad function
call.

Also, in the _cast function that handles dynamic_casts, the where clause uses
a bitwise AND, changed it to a short-circuiting AND. Really, just a cosmetic
change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14137 3a8e244f-b0f2-452b-bcba-4c88e055c3ca