Fix the message for out of memory allocating class
In constructors, the string passed to chpl_alloc to print if an error occurs
used the return type of the constructor to get the name of the class. Since
function return types aren't known until function resolution, the type of the
class would be printed as "_unknown". Instead, use the type of "this", which
is known when building the constructor.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14140 3a8e244f-b0f2-452b-bcba-4c88e055c3ca