improved param function support and fixed a bug
Fixed a bug in which param functions with multiple return statements
failed in a spectacular way--a compile-time constant would change its
value; for example, true might become false and 0 might become 1.
Added a couple of internal assertions to make sure this won't happen
in the future.  Made it so that when resolving a parameter function's
body, we finish after the first encountered return statement is
evaluated.  Since control flow must be already evaluated, there can
never be more than one evaluated.  Removed a future.

Added error checking for param functions that contain control flow
that is not evaluated at compilation time.  Update the spec and added
two test cases.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14154 3a8e244f-b0f2-452b-bcba-4c88e055c3ca