Added a number of unimplemented feature/bug futures. Specifically:
fillRandom - fillRandom() is a function described in the Random module,
described by the spec, to fill an array with random variables.  Actually
calling it causes a compiler error.

rankOnArrayOfArrays - Calling the rank field on a domain for an array of
arrays produces a compiler error.

returnArrayTuple - Specifying that a function should return a tuple of
array elements produces a compiler error.

initiailize2DArrayFromTuple - Initializing or assigning a 2-dimensional array
to a tuple of tuples (of the same size and dimensionality) will produce a
compiler error (the error message is a poor one as well).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14155 3a8e244f-b0f2-452b-bcba-4c88e055c3ca