I've now made a fix so that the three parallel tree operations, refine, compress, and reconstruct, should run with any numLocales and maxThreads settings without deadlocking (but with one exception as noted below).
With a single locale, the operations can execute with 1 or more
threads.

With multiple locales, maxThreads can have a value of 2 or more.
A multi-locale execution isn't currently doable with only 1 thread per
locale. I'm not sure if this can be improved in light of the remote
sync variable access pointed out by Sam.

The three codes now employ the foll simple rule to determine whether
or not a parent node should perform the corr operation on the child
nodes as parallel tasks:

if ( n+1 < log2(maxThreads) ) then
  begin/cobegin 
    on ..  
    on ..
  else {
    recursive_call (...);
    recursive_call (...);
  }

Here n denotes the level (or depth) of the parent node, n>=0.

This rule is explained by the binary tree traversal behavior of the
three operations, and accounts for the extreme case where all tasks
are spawned on the parent node's locale.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14159 3a8e244f-b0f2-452b-bcba-4c88e055c3ca