Fixed a bug with tuple-to-array assignment in which the multi-dimensional code was not working, but almost was.
Made unresolved call errors use square brackets if the call site did.

This fixes

  test/arrays/stonea/initialize2DArrayFromTuple.chpl

but I left the future because the test is an error.  The error message
looks acceptable to me in that I don't think we can do much better.
The issue here is that the element type of the array is specified
before the domain.  Swapping these fixes the test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14162 3a8e244f-b0f2-452b-bcba-4c88e055c3ca