improved support for explicit module accesses
Added support for explicit module uses on function calls.  This works
by marking the function call by passing it a moduleToken as its first
argument followed by the module.  This is then used for visible
function lookup.  This moduleToken is inserted during scope resolution
if the symbol is an overloadable function (one defined with
parentheses).  Then during function resolution, the module is used to
restrict the visible functions and to disambiguate function resolution
calls by visibility.

Refined the semantics of explicit module naming for all symbols
(including functions) so that the chosen symbol must be defined
top-level to the specified module (not outside of it or used from it).
Inserted error checking.  Updated the spec.

If we call a generic function via an explicit module name, then the
point of instantiation may not otherwise use that module.  This means
that code inside this function will not have access to this module.
Made it so that such an explicit module access inserts a 'use' around
this point of instantiation (via insertUseForExplicitModuleCalls).
Updated the spec.

Specialized function resolution errors for unresolved or ambiguous
calls where a module is specified.  The call is now printed with this
module.

Modified check for illegal module uses to allow modules to be used
after the moduleToken.

Duplicated David's test modules/diten/modulescopes.chpl since it is
now in error.  The new test modules/diten/modulescopes2.chpl moves the
variable into the innermost module and produces the original good.  I
left the old test to capture the error message.

Added a test of a call to a parenthesesless function with an explicit
module specified (modules/deitz/test_module_access_fn_no_parens.chpl).

Removed two futures related to the tests of error checking for symbols
that are in a scope outside of a module.  Added two new tests of this
same error checking for overloaded functions.

Removed Andy's future of calling a generic function after explicitly
specifying the Random module such that that call is the point of
instantiation.

Removed two of Brad's futures for calling functions after explicitly
specifying modules.

In total, removed five futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14172 3a8e244f-b0f2-452b-bcba-4c88e055c3ca