Add some tests and futures involving coercion and params
These tests are a result of looking at Andy's
domains/stonea/rankOnArrayOfArrays.chpl test before the int->string compile
time conversion was added. These overload a function for param and non-param
cases and attempt to pass parameter values that can be coerced at runtime,
but not at compile time. The non-param versions should be chosen since the
coercion needs to be done at runtime.

In coerce_param_imag_to_real.chpl, the "i" is omitted in the error message
about passing an imaginary param to a function expecting a real.

In param_int_to_complex2, the param formal is not used in the body of the
function which allows the call to succeed when it should be an error.

The enum->int and enum->string compile time conversions should be added.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14173 3a8e244f-b0f2-452b-bcba-4c88e055c3ca