enabled support for querying the eltType of an array
As in

  def foo(A: [D] ?t) { ... }

Refactored the array_type parsing code in chapel.y to distinguish
between generic and non-generic arrays at parse time.  Thus using such
syntax in variable declarations is now a syntax error.

Simplified and cleaned the code that handles array formals in
normalization, and extended it to the above support.

Removed the 9th oldest future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14174 3a8e244f-b0f2-452b-bcba-4c88e055c3ca