added support for user-defined classes for generic classes
In such a constructor, formals must be specified for each generic
field in the generic class unless the generic field is defined to have
a default value.  These are matched up and set by name.  If some
generic fields are not listed in the constructor, an error is flagged.

Fixed a .good error and removed the oldest future:
  reductions/bradc/manual/threeclasstypes-construct.future

Added two test cases for generic user-defined constructor and removed
a future for a test that now works.

Updated the spec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14176 3a8e244f-b0f2-452b-bcba-4c88e055c3ca