First step at trying Steve's leader/follower scheme
Here, I've taken my parallel version of BradsBlock1D.chpl,
stored in BradsBlock1DPar.chpl, and added a "newThese()"
function to the domain class that implements a leader/follower
scheme as proposed by Steve (which I'm still a bit on the
fence about, but don't yet have a reasonable counterproposal
for).

Created new test stream-block1dpar.chpl that uses this
BradsBlock1DPar distribution and rewrites the 
A = B + alpha * C statement to use explicit invocations
of the newThese() leader/follower iterator.  This seems
to divide up the work appropriately between a number of
tasks (debugging output can be seen by compiling using
-sdebugBradsBlock1D=true).  However, the on clause that
I was hoping would spread the computation onto the
appropriate locales does not seem to work -- probably
for good reasons like domains having value semantics and
being copied onto "here" in the iterator.  This raises
a bit of a puzzle, however, for where the on clause should
go and how it should work in the leader/follower scheme.


Next steps:

- come up with a proposal for this on clause mystery (I'd
  be happy for someone else to run with this, as I'm
  likely to be distracted by other stuff in the coming days)

- add a leader/follower iterator to the array itself -- my
  current rewrite uses random access of the array, which
  isn't ideal.  I will plan on working on this in the next
  few days.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14180 3a8e244f-b0f2-452b-bcba-4c88e055c3ca