Next step with leader/follower exploration
Added a leader/follower iterator for arrays like the one with
domains and made a second copy of my test that uses array
iteration rather than domain.  This one is slightly more
interesting because it uses a zippered follower iterator.
However, there are still a number of issues (like not using
multiple locales, doing too much random accessing, etc.)
These are captured in TODOs in the code.

One other surprise was that the follower iterator needs to be
a var iterator so that it yields references.  Yet the leader
iterator doesn't want to be because it yields a non-lvalue.
Steve and I have talked about possible ways to deal with this
including using overloading and where clauses to separate
the leader/follower into separate function definitions (which
I wanted to do stylistically anyway due to concerns about lack
of code reuse and arguments that didn't make sense in solo
and follower contexts).  That'll be my next step.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14182 3a8e244f-b0f2-452b-bcba-4c88e055c3ca