Fixed some bugs in the deadlock detection
When determining if a given thread is the last unblocked thread, the main thread was not taken into account.

Also, the time to wake up was assumed to be relative to the current time, when in fact, it needs to be an absolute time.  In addition, the time to wake up was stored in a global variable, so it was not thread safe.

When deadlock is detected, a message is now emitted to this effect.

Finally, a thread is marked as being potentially blocked only if the sync or single variable it is trying to access is actually not available.  I.e., it is marked as being potentially blocked only if it is actually blocked, even if for a small amount of time, whereas before it was marked as potentially blocked without checking to see if the sync or single variable was actually unavailable.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14184 3a8e244f-b0f2-452b-bcba-4c88e055c3ca