Refactored update_symbols to use the new traversal mechanism rather than collect_asts.
Refactored a number of uses of the copy method so that the optional
'map' argument is used only for symbols and types.  The only place
that relies on the map containing expressions is the iterator
transformation.  Collecting expressions adds measurable overhead to
compilation so this map should be simplified when the iterator
transformation is reimplemented.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14192 3a8e244f-b0f2-452b-bcba-4c88e055c3ca