Reverted chplsys.c so it now reports the full amount of physical memory instead of the maximum available via malloc in 32-bit programs on Mac OS X.
To get around the problem of not being able to run the HPCC benchmarks on Mac OS X systems with large amounts of physical memory, when on a 64-bit capable PowerPC-based machine, all Chapel programs are built as 64-bit executables.  On Intel-based Macs, they are still built as 32-bit executables, because I don't know how to distinguish 64-bit capable Intel processors, short of running a program or using the graphical user interface to determine what kind of processor it is.  (The alternative would be to build both 32- and 64-bit executables, at the expense of twice the compile time, and then have the OS run the favored version, which would be the 64-bit executable on 64-bit capable processors.)

Now, on my Mac at home, only two gdb-related regression tests fail.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14206 3a8e244f-b0f2-452b-bcba-4c88e055c3ca