Allow overriding with a method that is a subtype of the original method type
Instead of checking that the overriding method has the same return type as the
method it overrides, check that it is the same or a subtype.

When building up the dynamic dispatch checks at call sites, cast the result of
the call to the overridden method's type if it isn't already the same.

Make this behavior explicit in the spec.

When creating the arguments for fields in super classes in subclass default
constructors, add their names to the set of names not to reuse. This way a
name can be shadowed multiple times in the inheritance hierarchy.

Remove two .futures about overriding with a subtype.

Add a working test with a function returning both a class and a subclass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14210 3a8e244f-b0f2-452b-bcba-4c88e055c3ca