Add nearestMutualParentClass test.
A few interesting functions are defined in this test. First, isSubType, which
takes two class types and returns the param true if the first is a subclass of
the second, param false otherwise.

Next, getSuperType takes one class type and returns its super type.

Finally, nearestMutualParentClass takes a variable number of class types and
returns the type of the nearest class in the inheritance chain that can hold
values for all of the input types.

As Steve suggested might be true, no primitives are needed for any of this.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14212 3a8e244f-b0f2-452b-bcba-4c88e055c3ca