Made implementation of cobegin/coforall more similar to sync statment
The task list in the _EndCount class instance created for cobegin/coforall
is used instead of an independent task list, which used to be passed in to the
wrapper function as a dummy argument.  In order to figure out what task list
to use when generating C code, the compiler looks for a field in the bundled
argument called either "_ref(_EndCount)" or "_wide__ref__EndCount" (the actual
name depends on whether a wide ref is being used), and then passes in the
enclosed task list.

In order to use the same generated code for both cobegin and coforall
statements, I resorted to a kludge: I cloned the collect_asts() function and
special-cased it to look for index variables used in coforall statements, so
only those variables are put in the heap, instead of all symbols used in
coforall statements.  Ideally, this is how it should be; however, before, the
_EndCount instance that a coforall statement used was put on a heap, whereas the
_EndCount instance used in a cobegin statement was not on the heap, because none
of the symbols used in a cobegin are put on the heap.  Without this kludge,
different code would have to be generated to access the task list, depending on
whether it was associated with a cobegin or a coforall statement.

In the pthread runtime, now the only difference between how sync and cobegin/
coforall statements are handled is that, as before, when chpl_add_to_task_list()
is called to handle a begin statement, the given task is immediately added to
the task pool, whereas for cobegin/coforall statements, tasks are only added to
the task pool when chpl_process_task_list() is called.  The memory allocated for
the task list is no longer freed under any circumstances until
chpl_free_task_list() is called.  Also, chpl_process_task_list() no longer calls
chpl_execute_tasks_in_list() when dealing with a cobegin/coforall statement;
instead, it relies on _waitEndCount() to call the latter.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14218 3a8e244f-b0f2-452b-bcba-4c88e055c3ca