Cleaned up handling of gdb cleanup
For quite some time now, our --gdb flag on the compiler has not
cleaned up after itself if the program is re-run or terminated
prematurely.  This was due to a combination of a crufty
implementation on our part and seeming bugs/instabilities in
gdb.  A test of this could be done by compiling a program
with --gdb, setting a breakpoint in codegen, typing "run"
every time the breakpoint is hit, and watching 
/tmp/chpl-$user-$pid.deleteme directories accumulate.

With gdb 6.6, the instability in gdb is fixed, and so with
this change, I've re-enabled our cleaning up after ourselves
and also cleaned up our approach to doing so.  Now, when
a run or quit command is executed, the Chapel compiler
variable tmpdirname is inspected to see whether it is NULL
or not.  If it is not, gdb hook commands run our normal
cleanup_for_exit() call before running the user's command.

Previously, we were doing a similar check by hacking values
into the internal gdb variable $_exitcode to determine whether
the program was running or had terminated.  This approach
might also start working again in gdb 6.6 (I didn't check),
but isn't as attractive given how hacky it feels.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14221 3a8e244f-b0f2-452b-bcba-4c88e055c3ca