Fix some problems with sprintf calls
In a few places there were calls to sprintf with overlapping source and
destination locations. The newer valgrind we just started using caught this
issue. Instead of copying the string into itself when appending to it, just
append directly to the end of it.

In countTokens, also fixed a potential buffer overflow. It previously checked
for an overflow and issued a fatal error if one happened, but only after the
fact. Now, avoid the overflow and issue the error if it would have overflowed.

Also, when printing a one line comment for --print-code-size, use '//' instead
of '--'. Fix a test that expected '--' style comments.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14229 3a8e244f-b0f2-452b-bcba-4c88e055c3ca