added 'delete' (paired with Samuel)
Added keyword 'delete' and enabled parsing of a delete statement as in

  delete <expr> ;

Modified the parser to allow functions starting with ~ for use as
destructors.  All such functions are renamed "chpl_destroy".

The delete statement inserts a call to the CHPL_FREE primitive.  In
normalize, we insert a call to chpl_destroy before such a call.  To
do: This can probably be simplified so that we insert both of these
calls during parsing, but we still need to insert a temporary so that
we do not evaluate the expression being deleted twice.

In the modules, we defined chpl_destroy over object so that it will
work on any class even if the class does not define its own
destructor.  And we also defined chpl_destroy over everything else to
give an error that the user is trying to delete a value.

To do:

Error checking on improper declarations of destructors (declaring two
destructors and/or using ~ before the name of a method that is not the
class name or on a function.

Add array and domain tests of different functionality to make sure
that we free memory.

Insert calls to destructors and CHPL_FREE primitive to delete arrays
and domains after scope.

Issues: functions that return arrays, arrays that are put on the heap,
etc.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14230 3a8e244f-b0f2-452b-bcba-4c88e055c3ca