Next draft of leader/follower iterators in stream rewrite
In this draft, I've split the leader and follower iterators
into three separate calls, as motivated by my last draft
(captured in BradsBlock1DPar-deitzditen.chpl).  One is
for sequential iteration, the second a leader, and the
third a follower.  This approach strikes me as having
a few different advantages, including:

1) not requiring these three logically different behaviors
   to be lumped into a single routine

2) allowing the follower to be a var iterator without
   requiring the leader to be

3) allowing each to only take the arguments that it
   requires (in particular, the sequential and leader
   don't need the list of indices to follow)

4) allows the leader to be defined in an inlineable way.

As a result of this, I made the array follower iterator
into a var iterator which allowed me to rewrite the
block1dpar-arr version of the benchmark to zipper a, b,
and c without involving the domain.

Several TODOs remain, captured in the comments.  The
main next steps that I would like to explore are:

a) moving the parallelism and on clause into the leader

b) exploring whether local leader/follower parallelism
   can be cleanly put into the local domain/array classes.


As part of this change, I also added a README file to help
me remember what the existing drafts of the benchmark are
and why they are worthwhile.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14231 3a8e244f-b0f2-452b-bcba-4c88e055c3ca