Adding developer support for printing references
Added a(n undocumented) routine to ChapelBase.chpl that 
permits a developer to convert a Chapel reference into
a string for the purposes of debugging.  The routine
is named chpldev_revToString() and it takes an arbitrary
value as input and returns a string formatted with %p
in C.

I haven't done broad testing of this, and the one test
I'm checking in ensures that the routine can be called
more than anything.  However, I did find this invaluable
in debugging a distribution in which a series of iterators
and this functions were dropping a reference somewhere
along the way for reasons that I couldn't determine.
Modifying the generated code was my next best approach,
but would've taken a large effort and been very fragile.
With this routine, I could simply drop writeln()s into
all the functions and loops in question to make sure
that all the pointers matched.

Please send feedback if you use/try this, particularly
if you find problems with it.

The implementation of this was a little bit dicey --
see the comments preceding the routine for details.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14233 3a8e244f-b0f2-452b-bcba-4c88e055c3ca