Making Intel skip --gdb test on compiler
For some reason, the --gdb test on the compiler
sends our Intel compiles into an infinite loop.
Haven't had a chance to investigate yet, so am
disabling the test for Intel for the time being
until I get a chance to avoid filling up /tmp
again and again (it happened again last night).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14234 3a8e244f-b0f2-452b-bcba-4c88e055c3ca