Feature request: delete => nil
This is a feature request that when we delete a class reference
in Chapel, we set the reference back to nil to avoid some of the
most common errors.  Seems like we should be able to pull this
off without too much difficulty, but if not it would be nice
to know why.

(Obviously other references to this class instance could exist,
which would not be set to nil).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14238 3a8e244f-b0f2-452b-bcba-4c88e055c3ca