Fixing Intel --gdb issue
The Intel compiler seems to be more aggressive in removing
symbol names than the gnu compiler, so when running an
optimized version of the compiler with --gdb, we got
into troubles trying to query tmpdirname, which does not
exist.  This got gdb into a nice infinite loop in our
testing system.

This check-in fixes that problem by compiling the file
containing tmpdirname with the -g flag, whether or not
DEBUG is set during the make process.  To minimize the
amount of code affected by this decision, I moved the
variable into its own file -- tmpdirname.cpp.  This
seems to fix the problem for Intel compilers without
breaking the gnu compilers.

Removed the SKIPIF that I was using to avoid the
infinite loop over the weekend.

Future work: might also be useful to have a safeguard
way of quitting that disables the hook-quit command
before trying to execute the quit command.  Currently,
halt is also implemented in terms of quit, and can
therefore run into similar problems.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14240 3a8e244f-b0f2-452b-bcba-4c88e055c3ca