reworked tuple implementation
Moved the code to handle zipper iteration from the standard modules
into the compiler.  This means that when the compiler expands a
for-loop over an iterator it reasons about tuples and generates calls
to the iterator interface over multiple iterators for the single loop.

Simplified the iterator interface, removing the cursor.  For a loop

  for ind in iter() {
    ...
  }

the compiler stamps out the following code:

  var ic = iter(); // via _getIterator in the standard modules

  ic.advance();
  var cond = ic.hasMore();
  for (;cond;) {
    ind = ic.getValue();
    ...
    ic.advance();
    cond = ic.hasMore();
  }

The major change in this simplification is that fields are accessed
more directly in the advance() function than they were in the
getNextCursor() function.  Whereas we used to read the fields into
local variables at the beginning of this function (whether we needed
to or not (though in all cases only once)), we now read them as
needed.

In this process, disabled the single loop iterator optimization.  I
plan to reenable this.

Named the primitives used by halt, etc., in the modules to
PRIMITIVE_RT_ERROR and PRIMITIVE_RT_WARNING so they can be easily used
in the compiler.  The halts for non-equal zipper iteration have been
reworked so that the compiler generates them using these primitives
directly.

Gathered the module code related to iterators at the end of the
internal ChapelArray module.

Added functions iteratorIndex and iteratorIndexType to the modules in
order to determine the index type an iterator returns.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14241 3a8e244f-b0f2-452b-bcba-4c88e055c3ca