Improvements to --gdb flag on compiler
Quick review:
* In general, use "quit" and "run" within gdb to quit and run
  when using the --gdb flag -- we define hooks that cause the
  the compiler to clean up after itself if it's created a 
  temporary directory, so this is the recommended way to quit
  and run when debugging the compiler.

* Sometimes, things can get horked up within gdb to the extent
  that accessing memory (like tmpdirname, to see if it's set)
  doesn't work so well.  If this happens (i.e., if quit and
  run complain rather than doing what they're supposed to), then
  use "halt" or "rerun" to force the issue.  These commands
  will not clean up the temporary directory from the last run,
  so will cause your /tmp directory to get littered over time.
  If you find reproducible cases that require halt and rerun
  to be used, please let Brad know about them.

Steve ran into some cases where "halt" under the --gdb flag
was not working cleanly, requiring a suspend + kill in order
to exit gdb.  This was due in part to the fact that halt
calls 'quit' and seems to call the quit-hook as well (which
I don't think used to be the case?).

Rewrote the halt, rerun, and run/quit hooks so that halt and
rerun force a quit/run even if tmpdirname is not accessible,
using a gdb convenience variable to communicate between the
commands/hooks to indicate that the test of tmpdirname and
attempt to cleanup should be skipped.

Also restructured the way the gdb.commands file is created
so that the common suffix (the bulk of the file) is
concatenated from the (new) $CHPL_HOME/etc/gdb.commands file
rather than generated in the compiler using fprintfs.  This
should make it easier to experiment with alternate commands
in a lighter-weight manner (or allow developers to create
their own commands).

Orthogonally, added some standard svn:ignore files to some
test directories to clean up my svn status message.






git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14249 3a8e244f-b0f2-452b-bcba-4c88e055c3ca